Updated, normed and tested on 26/06/2019

+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: yu-lin <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/27 09:24:27 by yu-lin            #+#    #+#             */
-/*   Updated: 2019/06/20 14:11:49 by yu-lin           ###   ########.fr       */
+/*   Updated: 2019/06/26 09:39:18 by yu-lin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	unsigned long	h;
-	int				n;
+	size_t	h;
+	size_t	n;
 
 	h = 0;
 	if (needle[0] == '\0')
